Restored broken TrackView KeyUI controls
