+++ Code/Editor/TrackView/ScreenFaderKeyUIControls.cpp
@@ -14,77 +14,8 @@
 #include <CryCommon/Maestro/Types/AnimParamType.h>
 
 // Editor
+#include "KeyUIControls.h"
 #include "TrackViewKeyPropertiesDlg.h"
-
-
-//-----------------------------------------------------------------------------
-//!
-class CScreenFaderKeyUIControls
-    : public CTrackViewKeyUIControls
-{
-    CSmartVariableArray     mv_table;
-    CSmartVariable<float>   mv_fadeTime;
-    CSmartVariable<Vec3>    mv_fadeColor;
-    CSmartVariable<QString> mv_strTexture;
-    CSmartVariable<bool>    mv_bUseCurColor;
-    CSmartVariableEnum<int> mv_fadeType;
-    CSmartVariableEnum<int> mv_fadechangeType;
-
-public:
-    //-----------------------------------------------------------------------------
-    //!
-    bool SupportTrackType(const CAnimParamType& paramType, [[maybe_unused]] EAnimCurveType trackType, [[maybe_unused]] AnimValueType valueType) const override
-    {
-        return paramType == AnimParamType::ScreenFader;
-    }
-
-    //-----------------------------------------------------------------------------
-    //!
-    void OnCreateVars() override
-    {
-        AddVariable(mv_table, "Key Properties");
-
-        mv_fadeType->SetEnumList(nullptr);
-        mv_fadeType->AddEnumItem("FadeIn", IScreenFaderKey::eFT_FadeIn);
-        mv_fadeType->AddEnumItem("FadeOut", IScreenFaderKey::eFT_FadeOut);
-        AddVariable(mv_table, mv_fadeType, "Type");
-
-        mv_fadechangeType->SetEnumList(nullptr);
-        mv_fadechangeType->AddEnumItem("Linear", IScreenFaderKey::eFCT_Linear);
-        mv_fadechangeType->AddEnumItem("Square", IScreenFaderKey::eFCT_Square);
-        mv_fadechangeType->AddEnumItem("Cubic Square", IScreenFaderKey::eFCT_CubicSquare);
-        mv_fadechangeType->AddEnumItem("Square Root", IScreenFaderKey::eFCT_SquareRoot);
-        mv_fadechangeType->AddEnumItem("Sin", IScreenFaderKey::eFCT_Sin);
-        AddVariable(mv_table, mv_fadechangeType, "ChangeType");
-
-        AddVariable(mv_table, mv_fadeColor, "Color", IVariable::DT_COLOR);
-
-        mv_fadeTime->SetLimits(0.f, 100.f);
-        AddVariable(mv_table, mv_fadeTime, "Duration");
-        AddVariable(mv_table, mv_strTexture, "Texture", IVariable::DT_TEXTURE);
-        AddVariable(mv_table, mv_bUseCurColor, "Use Current Color");
-    }
-
-    //-----------------------------------------------------------------------------
-    //!
-    bool OnKeySelectionChange(CTrackViewKeyBundle& keys) override;
-
-    //-----------------------------------------------------------------------------
-    //!
-    void OnUIChange(IVariable* pVar, CTrackViewKeyBundle& keys) override;
-
-    unsigned int GetPriority() const override { return 1; }
-
-    static const GUID& GetClassID()
-    {
-        // {FBBC2407-C36B-45b2-9A54-0CF9CD3908FD}
-        static const GUID guid =
-        {
-            0xfbbc2407, 0xc36b, 0x45b2, { 0x9a, 0x54, 0xc, 0xf9, 0xcd, 0x39, 0x8, 0xfd }
-        };
-        return guid;
-    }
-};
 
 //-----------------------------------------------------------------------------
 bool CScreenFaderKeyUIControls::OnKeySelectionChange(CTrackViewKeyBundle& keys)
