+++ Code/Editor/TrackView/SoundKeyUIControls.cpp
@@ -13,51 +13,8 @@
 #include <CryCommon/Maestro/Types/AnimParamType.h>  // for AnimParamType
 
 // Editor
+#include "KeyUIControls.h"
 #include "TrackViewKeyPropertiesDlg.h"  // for CTrackViewKeyUIControls
-
-
-//////////////////////////////////////////////////////////////////////////
-class CSoundKeyUIControls
-    : public CTrackViewKeyUIControls
-{
-public:
-    CSmartVariableArray mv_table;
-    CSmartVariableArray mv_options;
-
-    CSmartVariable<QString> mv_startTrigger;
-    CSmartVariable<QString> mv_stopTrigger;
-    CSmartVariable<float> mv_duration;
-    CSmartVariable<Vec3> mv_customColor;
-
-    void OnCreateVars() override
-    {
-        AddVariable(mv_table, "Key Properties");
-        AddVariable(mv_table, mv_startTrigger, "StartTrigger", IVariable::DT_AUDIO_TRIGGER);
-        AddVariable(mv_table, mv_stopTrigger, "StopTrigger", IVariable::DT_AUDIO_TRIGGER);
-        AddVariable(mv_table, mv_duration, "Duration");
-        AddVariable(mv_options, "Options");
-        AddVariable(mv_options, mv_customColor, "Custom Color", IVariable::DT_COLOR);
-    }
-    bool SupportTrackType(const CAnimParamType& paramType, [[maybe_unused]] EAnimCurveType trackType, [[maybe_unused]] AnimValueType valueType) const override
-    {
-        return paramType == AnimParamType::Sound;
-    }
-    bool OnKeySelectionChange(CTrackViewKeyBundle& selectedKeys) override;
-    void OnUIChange(IVariable* pVar, CTrackViewKeyBundle& selectedKeys) override;
-
-    unsigned int GetPriority() const override { return 1; }
-
-    static const GUID& GetClassID()
-    {
-        // {AB2226E5-D593-49d2-B7CB-989412CAAEDE}
-        static const GUID guid =
-        {
-            0xab2226e5, 0xd593, 0x49d2, { 0xb7, 0xcb, 0x98, 0x94, 0x12, 0xca, 0xae, 0xde }
-        };
-        return guid;
-    }
-};
-
 
 //////////////////////////////////////////////////////////////////////////
 bool CSoundKeyUIControls::OnKeySelectionChange(CTrackViewKeyBundle& selectedKeys)
